paths: Split find_mtimes() out of module
