+++ mopidy/internal/path.py
@@ -3,12 +3,13 @@ from __future__ import absolute_import, unicode_literals
 import logging
 import os
 import re
-import stat
-import threading
 
-from mopidy import compat, exceptions
-from mopidy.compat import pathlib, queue, urllib
-from mopidy.internal import encoding, xdg
+
+from mopidy import compat
+from mopidy.compat import pathlib, urllib
+from mopidy.internal import xdg
+# Reexport in old location for Mopidy-Local's use
+from mopidy.internal.mtimes import find_mtimes  # noqa
 
 
 logger = logging.getLogger(__name__)
@@ -87,98 +88,6 @@ def expand_path(path):
     if '$' in path:
         return None
     return pathlib.Path(path).expanduser().resolve()
-
-
-def _find_worker(relative, follow, done, work, results, errors):
-    """Worker thread for collecting stat() results.
-
-    :param str relative: directory to make results relative to
-    :param bool follow: if symlinks should be followed
-    :param threading.Event done: event indicating that all work has been done
-    :param queue.Queue work: queue of paths to process
-    :param dict results: shared dictionary for storing all the stat() results
-    :param dict errors: shared dictionary for storing any per path errors
-    """
-    while not done.is_set():
-        try:
-            entry, parents = work.get(block=False)
-        except queue.Empty:
-            continue
-
-        if relative:
-            path = os.path.relpath(entry, relative)
-        else:
-            path = entry
-
-        try:
-            if follow:
-                st = os.stat(entry)
-            else:
-                st = os.lstat(entry)
-
-            if (st.st_dev, st.st_ino) in parents:
-                errors[path] = exceptions.FindError('Sym/hardlink loop found.')
-                continue
-
-            parents = parents + [(st.st_dev, st.st_ino)]
-            if stat.S_ISDIR(st.st_mode):
-                for e in os.listdir(entry):
-                    work.put((os.path.join(entry, e), parents))
-            elif stat.S_ISREG(st.st_mode):
-                results[path] = st
-            elif stat.S_ISLNK(st.st_mode):
-                errors[path] = exceptions.FindError('Not following symlinks.')
-            else:
-                errors[path] = exceptions.FindError('Not a file or directory.')
-
-        except OSError as e:
-            errors[path] = exceptions.FindError(
-                encoding.locale_decode(e.strerror), e.errno)
-        finally:
-            work.task_done()
-
-
-def _find(root, thread_count=10, relative=False, follow=False):
-    """Threaded find implementation that provides stat results for files.
-
-    Tries to protect against sym/hardlink loops by keeping an eye on parent
-    (st_dev, st_ino) pairs.
-
-    :param str root: root directory to search from, may not be a file
-    :param int thread_count: number of workers to use, mainly useful to
-        mitigate network lag when scanning on NFS etc.
-    :param bool relative: if results should be relative to root or absolute
-    :param bool follow: if symlinks should be followed
-    """
-    threads = []
-    results = {}
-    errors = {}
-    done = threading.Event()
-    work = queue.Queue()
-    work.put((os.path.abspath(root), []))
-
-    if not relative:
-        root = None
-
-    args = (root, follow, done, work, results, errors)
-    for i in range(thread_count):
-        t = threading.Thread(target=_find_worker, args=args)
-        t.daemon = True
-        t.start()
-        threads.append(t)
-
-    work.join()
-    done.set()
-    for t in threads:
-        t.join()
-    return results, errors
-
-
-def find_mtimes(root, follow=False):
-    results, errors = _find(root, relative=False, follow=follow)
-    # return the mtimes as integer milliseconds
-    mtimes = {f: int(st.st_mtime * 1000) for f, st in results.items()}
-    return mtimes, errors
 
 
 def is_path_inside_base_dir(path, base_path):
